carve out sinks submodule
