+++ capture/src/lib.rs
@@ -8,7 +8,7 @@ pub mod prometheus;
 pub mod redis;
 pub mod router;
 pub mod server;
-pub mod sink;
+pub mod sinks;
 pub mod time;
 pub mod token;
 pub mod utils;
