+++ capture/src/router.rs
@@ -11,14 +11,14 @@ use tower_http::trace::TraceLayer;
 
 use crate::health::HealthRegistry;
 use crate::{
-    capture, limiters::billing_limits::BillingLimiter, redis::Client, sink, time::TimeSource,
+    capture, limiters::billing_limits::BillingLimiter, redis::Client, sinks, time::TimeSource,
 };
 
 use crate::prometheus::{setup_metrics_recorder, track_metrics};
 
 #[derive(Clone)]
 pub struct State {
-    pub sink: Arc<dyn sink::EventSink + Send + Sync>,
+    pub sink: Arc<dyn sinks::EventSink + Send + Sync>,
     pub timesource: Arc<dyn TimeSource + Send + Sync>,
     pub redis: Arc<dyn Client + Send + Sync>,
     pub billing: BillingLimiter,
@@ -30,7 +30,7 @@ async fn index() -> &'static str {
 
 pub fn router<
     TZ: TimeSource + Send + Sync + 'static,
-    S: sink::EventSink + Send + Sync + 'static,
+    S: sinks::EventSink + Send + Sync + 'static,
     R: Client + Send + Sync + 'static,
 >(
     timesource: TZ,
